Moved device history functionality into its own class.
